+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `material.MaterialTechnique` and `material.MaterialTechniqueFlags`.
 * Added fields `material.OutputAssignment.xyz` and `material.OutputAssignments.exprs_xyz`.
 * Added field `channel` to `shader_database.OutputExprFuncXyz`.
-* Added field` unk4_legacy` to `material.Material`.
+* Added fields `unk4_legacy`, `unk_param1`, `unk_param2` to `material.Material`.
 
 ### Changed
 * Renamed `shader_database.Dependency` to `shader_database.Value` and updated associated type names.